# openapi.yaml
openapi: 3.1.0
info:
  title: Autonomous AI Agent Safety & Permission Control System
  description: Local middleware firewall intercepting agent actions, scoring risk,
    and enforcing policy.
  version: 1.0.0
paths:
  /api/v1/operator/login:
    post:
      tags:
      - Operator
      summary: Login Operator
      operationId: login_operator_api_v1_operator_login_post
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Body_login_operator_api_v1_operator_login_post'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperatorLoginResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/operator/logout:
    post:
      tags:
      - Operator
      summary: Logout Operator
      operationId: logout_operator_api_v1_operator_logout_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - OAuth2PasswordBearer: []
  /api/v1/operator/me:
    get:
      tags:
      - Operator
      summary: Get Operator Me
      operationId: get_operator_me_api_v1_operator_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - OAuth2PasswordBearer: []
...
  /api/v1/ask:
    post:
      tags:
      - Question Answering & Placement
      summary: Ask a question over AMYPO database & institutional knowledge
      description: Natural language Q&A with mandatory in-line AURA safety evaluation
        and zero-hallucination grounding.
      operationId: ask_question_api_v1_ask_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AskRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AskResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/placement/match:
    get:
      tags:
      - Question Answering & Placement
      summary: Evaluate placement eligibility with exact mathematical formula
      description: 'Formula: Eligibility = (Normalized GPA * 0.40) + (Skills Match
        % * 0.60). Recommends gap courses if < 60%.'
      operationId: match_placement_api_v1_placement_match_get
      parameters:
      - name: student_id
        in: query
        required: true
        schema:
          type: string
          description: Student ID e.g. STU001
          title: Student Id
        description: Student ID e.g. STU001
      - name: company_id
        in: query
        required: true
        schema:
          type: string
          description: Company ID e.g. CMP001
          title: Company Id
        description: Company ID e.g. CMP001
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlacementMatchResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    AskRequest:
      properties:
        question:
          type: string
          title: Question
          description: Natural language question from student or staff
        user_id:
          anyOf:
          - type: string
          - type: 'null'
          title: User Id
          description: Optional user or agent identifier
      type: object
      required:
      - question
      title: AskRequest
    AskResponse:
      properties:
        answer:
          type: string
          title: Answer
          description: Grounded natural language answer
        sources:
          items:
            $ref: '#/components/schemas/SourceItem'
          type: array
          title: Sources
          description: Strict source citations with record_id and snippet
        confidence:
          type: number
          title: Confidence
          description: Cosine similarity confidence score between 0.0 and 1.0
      type: object
      required:
      - answer
      - sources
      - confidence
      title: AskResponse
...
    Body_login_operator_api_v1_operator_login_post:
      properties:
        grant_type:
          anyOf:
          - type: string
            pattern: password
          - type: 'null'
          title: Grant Type
        username:
          type: string
          title: Username
        password:
          type: string
          title: Password
        scope:
          type: string
          title: Scope
          default: ''
        client_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Client Id
        client_secret:
          anyOf:
          - type: string
          - type: 'null'
          title: Client Secret
      type: object
      required:
      - username
      - password
      title: Body_login_operator_api_v1_operator_login_post
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    OperatorLoginResponse:
      properties:
        access_token:
          type: string
          title: Access Token
        token_type:
          type: string
          title: Token Type
          default: bearer
        role:
          type: string
          title: Role
      type: object
      required:
      - access_token
      - role
      title: OperatorLoginResponse
    PlacementMatchResponse:
      properties:
        student_id:
          type: string
          title: Student Id
        student_name:
          type: string
          title: Student Name
        student_gpa:
          type: number
          title: Student Gpa
        student_attendance:
          type: number
          title: Student Attendance
        company_id:
          type: string
          title: Company Id
        company_name:
          type: string
          title: Company Name
        target_role:
          type: string
          title: Target Role
        package_lpa:
          type: number
          title: Package Lpa
        normalized_gpa:
          type: number
          title: Normalized Gpa
        skills_match_pct:
          type: number
          title: Skills Match Pct
        eligibility_score:
          type: number
          title: Eligibility Score
        is_eligible:
          type: boolean
          title: Is Eligible
        meets_gpa_cutoff:
          type: boolean
          title: Meets Gpa Cutoff
        meets_attendance_cutoff:
          type: boolean
          title: Meets Attendance Cutoff
        matched_skills:
          items:
            type: string
          type: array
          title: Matched Skills
        missing_skills:
          items:
            type: string
          type: array
          title: Missing Skills
        gap_courses:
          items:
            type: object
          type: array
          title: Gap Courses
      type: object
      required:
      - student_id
      - student_name
      - student_gpa
      - student_attendance
      - company_id
      - company_name
      - target_role
      - package_lpa
      - normalized_gpa
      - skills_match_pct
      - eligibility_score
      - is_eligible
      - meets_gpa_cutoff
      - meets_attendance_cutoff
      - matched_skills
      - missing_skills
      - gap_courses
      title: PlacementMatchResponse
    SourceItem:
      properties:
        record_id:
          type: string
          title: Record Id
          description: Unique identifier of the cited source record
        snippet:
          type: string
          title: Snippet
          description: Verbatim extracted snippet from the source document
      type: object
      required:
      - record_id
      - snippet
      title: SourceItem
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/operator/login